@@ -1,36 +1,25 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="EFCore.NamingConventions" Version="5.0.0" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0">
-          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        <PackageReference Include="EFCore.NamingConventions" Version="6.0.0" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.4" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="6.0.4">
           <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="5.0.10" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
+        <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="6.0.3" />
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
     </ItemGroup>
 
     <ItemGroup>
       <ProjectReference Include="..\eAkademik.Model\eAkademik.Model.csproj" />
       <ProjectReference Include="..\eAkademik.Seed\eAkademik.Seed.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Compile Remove="Migrations\20220404182503_CreateUserTable.Designer.cs" />
-      <Compile Remove="Migrations\ContextModelSnapshot.cs" />
-      <Compile Remove="Migrations\20220404183054_CreateUserTable.cs" />
-      <Compile Remove="Migrations\20220404183054_CreateUserTable.Designer.cs" />
-      <Compile Remove="Migrations\20220404184704_CreateUserTable.Designer.cs" />
-      <Compile Remove="Migrations\20220404184847_CreateUserTable.Designer.cs" />
-      <Compile Remove="Migrations\20220404185019_CreateUserTable.Designer.cs" />
-      <Compile Remove="Migrations\20220404185427_CreateUserTable.Designer.cs" />
-      <Compile Remove="Migrations\20220404185708_CreateUserTable.Designer.cs" />
-      <Compile Remove="Migrations\20220404185851_CreateUserTable.Designer.cs" />
-      <Compile Remove="Migrations\20220404185946_CreateUserTable.Designer.cs" />
     </ItemGroup>
 
     <ItemGroup>
