@@ -1,14 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using eAkademik.Model;
+﻿using eAkademik.Model;
 
-namespace eAkademik.API.Services
+namespace eAkademik.API.Services;
+
+public interface IUserService
 {
-    public interface IUserService
-    {
-        public Task<List<User>> GetUsers();
+    public Task<List<User>> GetUsers();
 
-        public Task<User> GetUser(Guid id);
-    }
+    public Task<User> GetUser(Guid id);
 }
