@@ -1,18 +1,18 @@
-﻿using System;
-using eAkademik.Model.Shared;
+﻿using eAkademik.Model.Shared;
 
-namespace eAkademik.Model
+namespace eAkademik.Model;
+
+public class User: IEntity
 {
-    public class User: IEntity
-    {
-        public Guid Id { get; set; }
+    public Guid Id { get; set; }
 
-        public string FirstName { get; set; }
+    public string FirstName { get; set; }
 
-        public string Email { get; set; }
+    public string LastName { get; set; }
 
-        public string LastName { get; set; }
+    public string Email { get; set; }
 
-        public string Password { get; set; }
-    }
+    public string Password { get; set; }
+
+    public bool IsDeleted { get; set; }
 }
